Export layer with mode "feature-metric".
Split the export into two helpers classes
to make it more readable.
By the way, remove a duplicated
KeywordColor declaration and move it
into a "models.ts" file.
