+++ src/app/modules/map-config/components/dialog-color-table/dialog-color-table.component.spec.ts
@@ -1,10 +1,9 @@
 import { DialogColorTableComponent } from './dialog-color-table.component';
-import { Spectator, createComponentFactory, mockProvider } from '@ngneat/spectator';
+import { Spectator, createComponentFactory } from '@ngneat/spectator';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MockComponent } from 'ng-mocks';
 import { ColorPickerWrapperComponent } from '@shared-components/color-picker-wrapper/color-picker-wrapper.component';
 import { ArlasCollaborativesearchService, ArlasColorGeneratorLoader } from 'arlas-wui-toolkit';
-import { HttpClient } from '@angular/common/http';
 import { CollectionService } from '@services/collection-service/collection.service';
 
 describe('DialogColorTableComponent', () => {
